requirements.txt
apache-airflow==2.8.4
psycopg2-binary
